testing/requirements.txt
click==8.1.3
pandas==1.5.1
plotly==5.10.0
PyYAML==6.0
kaleido==0.2.1
pydantic==1.10.2
requests==2.28.1
